:recycle: Move timer handling to timer facade
