+++ package.json
@@ -26,7 +26,7 @@
     "@essential-projects/iam_contracts": "3.1.0",
     "@essential-projects/timing_contracts": "3.0.0",
     "@process-engine/consumer_api_contracts": "0.16.1",
-    "@process-engine/process_engine_contracts": "26.0.0",
+    "@process-engine/process_engine_contracts": "feature~add_timer_facade_interface",
     "@types/clone": "0.1.30",
     "addict-ioc": "2.3.7",
     "bluebird": "3.5.1",
