+++ src/runtime/engine/handler/intermediate_event_handlers/intermediate_timer_catch_event_handler.ts
@@ -1,55 +1,36 @@
-import * as moment from 'moment';
-import * as uuid from 'uuid';
-
-import {ITimerService, TimerRule} from '@essential-projects/timing_contracts';
-
 import {
   IExecutionContextFacade,
   IFlowNodeInstanceService,
   IProcessModelFacade,
   IProcessTokenFacade,
+  ITimerFacade,
   Model,
   NextFlowNodeInfo,
   Runtime,
   TimerDefinitionType,
 } from '@process-engine/process_engine_contracts';
 
-import {IEventAggregator, ISubscription} from '@essential-projects/event_aggregator_contracts';
+import {ISubscription} from '@essential-projects/event_aggregator_contracts';
 
 import {FlowNodeHandler} from '../index';
 
-enum TimerBpmnType {
-  Duration = 'bpmn:timeDuration',
-  Cycle = 'bpmn:timeCycle',
-  Date = 'bpmn:timeDate',
-}
-
 export class IntermediateTimerCatchEventHandler extends FlowNodeHandler<Model.Events.IntermediateCatchEvent> {
 
-  private _eventAggregator: IEventAggregator;
   private _flowNodeInstanceService: IFlowNodeInstanceService = undefined;
-  private _timerService: ITimerService = undefined;
+  private _timerFacade: ITimerFacade;
 
-  constructor(eventAggregator: IEventAggregator,
-              flowNodeInstanceService: IFlowNodeInstanceService,
-              timerService: ITimerService) {
-
+  constructor(flowNodeInstanceService: IFlowNodeInstanceService, timerFacade: ITimerFacade) {
     super();
-    this._eventAggregator = eventAggregator;
     this._flowNodeInstanceService = flowNodeInstanceService;
-    this._timerService = timerService;
-  }
-
-  private get eventAggregator(): IEventAggregator {
-    return this._eventAggregator;
+    this._timerFacade = timerFacade;
   }
 
   private get flowNodeInstanceService(): IFlowNodeInstanceService {
     return this._flowNodeInstanceService;
   }
 
-  private get timerService(): ITimerService {
-    return this._timerService;
+  private get timerFacade(): ITimerFacade {
+    return this._timerFacade;
   }
 
   protected async executeInternally(flowNode: Model.Events.IntermediateCatchEvent,
@@ -65,8 +46,8 @@ export class IntermediateTimerCatchEventHandler extends FlowNodeHandler<Model.Ev
 
       let timerSubscription: ISubscription;
 
-      const timerType: TimerDefinitionType = this._parseTimerDefinitionType(flowNode.timerEventDefinition);
-      const timerValue: string = this._parseTimerDefinitionValue(flowNode.timerEventDefinition);
+      const timerType: TimerDefinitionType = this.timerFacade.parseTimerDefinitionType(flowNode.timerEventDefinition);
+      const timerValue: string = this.timerFacade.parseTimerDefinitionValue(flowNode.timerEventDefinition);
 
       const nextFlowNodeInfo: Model.Base.FlowNode = processModelFacade.getNextFlowNodeFor(flowNode);
 
@@ -85,108 +66,7 @@ export class IntermediateTimerCatchEventHandler extends FlowNodeHandler<Model.Ev
         resolve(new NextFlowNodeInfo(nextFlowNodeInfo, token, processTokenFacade));
       };
 
-      timerSubscription = await this._initializeTimer(flowNode, timerType, timerValue, timerElapsed);
+      timerSubscription = await this.timerFacade.initializeTimer(flowNode, timerType, timerValue, timerElapsed);
     });
-  }
-
-  private async _initializeTimer(flowNode: Model.Events.IntermediateCatchEvent,
-                                 timerType: TimerDefinitionType,
-                                 timerValue: string,
-                                 timerCallback: Function): Promise<ISubscription> {
-
-    const callbackEventName: string = `${flowNode.id}_${uuid.v4()}`;
-
-    switch (timerType) {
-      case TimerDefinitionType.cycle:
-        return this._startCycleTimer(timerValue, timerCallback, callbackEventName);
-      case TimerDefinitionType.date:
-        return this._startDateTimer(timerValue, timerCallback, callbackEventName);
-      case TimerDefinitionType.duration:
-        return this._startDurationTimer(timerValue, timerCallback, callbackEventName);
-      default:
-    }
-  }
-
-  private _parseTimerDefinitionType(eventDefinition: any): TimerDefinitionType {
-    if (eventDefinition[TimerBpmnType.Duration]) {
-      return TimerDefinitionType.duration;
-    }
-    if (eventDefinition[TimerBpmnType.Cycle]) {
-      return TimerDefinitionType.cycle;
-    }
-    if (eventDefinition[TimerBpmnType.Date]) {
-      return TimerDefinitionType.date;
-    }
-
-    return undefined;
-  }
-
-  private _parseTimerDefinitionValue(eventDefinition: any): string {
-    if (eventDefinition[TimerBpmnType.Duration]) {
-      return eventDefinition[TimerBpmnType.Duration]._;
-    }
-    if (eventDefinition[TimerBpmnType.Cycle]) {
-      return eventDefinition[TimerBpmnType.Cycle]._;
-    }
-    if (eventDefinition[TimerBpmnType.Date]) {
-      return eventDefinition[TimerBpmnType.Date]._;
-    }
-
-    return undefined;
-  }
-
-  private async _startCycleTimer(timerDefinition: string,
-                                 timerCallback: Function,
-                                 callbackEventName: string): Promise<ISubscription> {
-
-    const duration: moment.Duration = moment.duration(timerDefinition);
-
-    const timingRule: TimerRule = {
-      year: duration.years(),
-      month: duration.months(),
-      date: duration.days(),
-      hour: duration.hours(),
-      minute: duration.minutes(),
-      second: duration.seconds(),
-    };
-
-    const subscription: ISubscription = this.eventAggregator.subscribe(callbackEventName, () => {
-      timerCallback();
-    });
-
-    await this.timerService.periodic(timingRule, callbackEventName);
-
-    return subscription;
-  }
-
-  private async _startDurationTimer(timerDefinition: string,
-                                    timerCallback: Function,
-                                    callbackEventName: string): Promise<ISubscription> {
-
-    const duration: moment.Duration = moment.duration(timerDefinition);
-    const date: moment.Moment = moment().add(duration);
-
-    const subscription: ISubscription = this.eventAggregator.subscribeOnce(callbackEventName, () => {
-      timerCallback();
-    });
-
-    await this.timerService.once(date, callbackEventName);
-
-    return subscription;
-  }
-
-  private async _startDateTimer(timerDefinition: string,
-                                timerCallback: Function,
-                                callbackEventName: string): Promise<ISubscription> {
-
-    const date: moment.Moment = moment(timerDefinition);
-
-    const subscription: ISubscription = this.eventAggregator.subscribeOnce(callbackEventName, () => {
-      timerCallback();
-    });
-
-    await this.timerService.once(date, callbackEventName);
-
-    return subscription;
   }
 }
