+++ src/runtime/engine/index.ts
@@ -6,5 +6,6 @@ export * from './process_model_facade_factory';
 export * from './process_token_facade';
 export * from './process_token_facade_factory';
 export * from './sub_process_model_facade';
+export * from './timer_facade';
 
 export * from './handler/index';
